// front/sistema-tickets/src/app/models/personal.model.ts
export interface PersonalModelTs {
    /** Variabe para la clave del personal */
    cvePersonal: string;

    /** Variable para la clave del usuario */
    cveUsuario: string;
    
    /** Variable que contiene la clave del Rol */
    cveRol: string;

    /** Variable que contiene la clave del Rol */
    nombreRol: string;

    /** Variable que contiene el usuario */
    usuario: string;

    /** Variable que contiene el password del usuario */
    password: string;

    /** Variable que contiene el estatus del usuario */
    estatUsr: string;

    /** Variable que contiene el estatus de la cuenta */
    estatus: string;

    /** Variable que contiene el nombre de la persona */
    nombre: string;
}